add print errors kata

+++ codewars/src/Alphabet Sequence.kt
@@ -1,3 +1,3 @@
 fun alphaSeq(str: String) = str.toLowerCase().asIterable().sorted().joinToString(",") { c ->
-    c.toUpperCase() + c.toString().repeat(c.toInt() - 'a'.toInt())
+    c.toUpperCase() + c.toString().repeat(c - 'a')
 }
